simplify AbstractRpcRemotingClient.
